refactor(weave): extract snapshot_mutable_containers to own module
Why: weave_client.py was already large; the snapshot helper is a
self-contained spine-rebuilder that doesn't need to live next to the
client. Moves the function (and its docstring) to weave/trace/output_snapshot.py
and drops the leading underscore since it's now the module's public surface.
